Fixed Compilation Errors
Can now compile.

Next step is to draw something.

+++ VFS/VFS/Graphics.h
@@ -10,8 +10,8 @@ public:
 	void Shutdown();
 
 	auto GetWindow()->GLFWwindow* { return window_; };
-	int GetWindowWidth() const { return window_width_; }
-	int GetWindowHeight() const { return window_height_; }
+	int GetWindowWidth() const { return static_cast<int>(window_width_); }
+	int GetWindowHeight() const { return static_cast<int>(window_height_); }
 
 	void SetWindowSize(float width, float height) { window_width_ = width, window_height_ = height; }
